refactor: keep apiCall blocklist in executor.go
Move the HTTP blocklist helpers into executor.go so the fix does not
add extra files. Behavior is unchanged.
